robot_ufla/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
comtypes==1.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
PyAudio==0.2.13
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.66.1
urllib3==2.0.5
yarl==1.9.2
